fix: prevent null exception caused by integer division by zero
(cherry picked from commit eb1d1507346d225dea5fab86f0a3205760c65b6a)

+++ common/src/main/java/me/adda/terramath/math/functions/MathExtensions.java
@@ -103,7 +103,7 @@ public class MathExtensions {
         return mean + stddev * random.nextGaussian();
     }
 
-    public static int gcd(Number a, Number b) {
+    public static double gcd(Number a, Number b) {
         int aVal = a.intValue();
         int bVal = b.intValue();
 
@@ -115,7 +115,7 @@ public class MathExtensions {
         return Math.abs(aVal);
     }
 
-    public static int lcm(Number a, Number b) {
+    public static double lcm(Number a, Number b) {
         int aVal = a.intValue();
         int bVal = b.intValue();
